Initial commit : data_preprocessing implementation
Why: Worked in local environment to implement data preprocessing.
This data preprocessing pipeline is used to prepare the data for the
training of the model. It includes the following steps:
1. Remove the scale bar from the images.
2. Organize the data into train, validation and test sets.
3. Data augmentation
4. YAML file generation for the training of the model.
5. Logging using WandB.

How:
Since this is the first commit and I worked in local without Github
properly setup, this would be too long to review everything included
here. Please refer to the README.md file for a detailed description
of the data preprocessing pipeline.

Tags: #data_preprocessing #data_augmentation #remove_scalebar

Committer: Hugo Walter <hugo.walter@inrae.fr>