Add few utils scripts
Why: These scripts comes handy to verify things. check_imgs.py is used
to check if the images are corrupted or not. clean_unwanted_classes.sh
is used to remove unwanted classes from the dataset. pytorch_cuda.py
is used to check if the GPU is available or not. validate_yaml.py is
used to validate the yaml file.

Tags: utils